algorithms->binarySearch.c: return a vector instead of a pointer
Previously, find_col_peak returned a pointer to a max_vector instead of returning
the vector by value. This means we unnecessarily malloc a vector, and because this
this a recursive function we have to worry about freeing the memory in the right
places at the right times.

Furthermore, we have to do some pointer house keeping e.g. ensuring malloc
correctly returns a pointer to the assigned memory location.

This approach allows us to make use of the stack and not have to worry about
freeing any memory during the recursive calls.